Let a checker read what only a reviewer had
Error Prone and NullAway, on `src/main` and blocking. Seventeen findings across
the codebase, and the interesting part is how few were style:

  - HarfBuzz.failure was dead. -Xlint does not see an unused private method.
  - GoldberryRuntime declared an IdentityHashMap as a Map, which hides that
    lookups are by reference -- the one property that field depends on.
  - PointerRouter used `instanceof Element` three times as a null check on a
    method already returning Element, so the type read as though it were in
    question when only presence ever was.
  - Font widened a 26.6 fixed-point long to double silently.
  - Showcase's background job swallowed its own failures: a `thenAccept` whose
    future nobody holds leaves the bar reading "reading the map…" for ever with
    nothing in the log. It has somewhere to go now.
  - TourState carried a javadoc reference split across two lines, `[` on one and
    the target on the next. Two independent Java parsers crash on that file --
    Error Prone's and Spotless's -- and neither said why.

The suppressions are all argued at their site. ReferenceEquality is off
globally: `Thread.currentThread() != owner` is the confinement check eighteen
files make and is correct every time, so the check has no signal here.
:natives has Error Prone off entirely -- it is hand-written FFM bindings, and
the plugin throws rather than reports on Blend2dContext.

NullAway runs in OnlyNullMarked mode, not AnnotatedPackages. The first attempt
declared the whole tree annotated, which reports four errors a file in code
nobody has annotated yet; this way a package is checked from the moment it opts
in and is invisible until then, which is how a codebase this size adopts it at
all. No package has opted in yet -- that sweep is its own commit.

-Werror is now conditional on -Pgoldberry.lenient, for triage only: adding a
checker reports findings one compilation unit at a time, and enumerating the
tail means letting the compile finish.

+++ core/src/main/java/io/github/digitalsmile/goldberry/text/font/Font.java
@@ -221,7 +221,10 @@ public final class Font implements AutoCloseable {
     public double widthOf(GlyphRun run) {
         requireUsable();
         Objects.requireNonNull(run, "run");
-        return toLogical(run.totalXAdvance());
+        // The cast is explicit because the advance is a 26.6 fixed-point `long`
+        // and `toLogical` takes a double: the widening was happening anyway, and
+        // saying so is the difference between a conversion and an accident.
+        return toLogical((double) run.totalXAdvance());
     }
 
     /// How wide `text` is once shaped, in logical units.
